btif_config: Fix pointer to freed memory
We change the scope of our 'string' declaration so the contents
are still in scope when we use a pointer to them.  Previously,
the contents were out of scope when we referenced them through
a pointer.

Bug: 148758680
Test: TreeHugger
Change-Id: Ib0aa689bf4547d079b77965d958cc382ba0276e6
Merged-In: Ib0aa689bf4547d079b77965d958cc382ba0276e6
(cherry picked from commit 3079ddcd325127ca9314e614603c7312dffdf3d4)